Rehauled probability system
Switched up the probability system so it uses actual probabilities instead of fixed values now. Much easier to manipulate. Also re-structured the files to reduce the number of lines on main.cpp.
